omp speedup

+++ src/benchmark.c
@@ -13,12 +13,12 @@ void bt_init(bt_Time *bt) {
 
 void bt_print(bt_Time *bt) {
 
-  printf("time_main: %lf s\n", bt->main / CLOCKS_PER_SEC);
-  printf("time_pdf: %lf %%\n", bt->pdf / bt->main * 100.);
-  printf("time_pdf_mupdf: %lf %%\n", bt->pdf_mupdf / bt->main * 100.);
-  printf("time_pdf_findplots: %lf %%\n", bt->pdf_findplots / bt->main * 100.);
-  printf("time_pdf_loopplots: %lf %%\n", bt->pdf_loopplots / bt->main * 100.);
-  printf("time_pdf_dct: %lf %%\n", bt->pdf_dct / bt->main * 100.);
+  printf("time_main: %f s\n", bt->main / ((double) CLOCKS_PER_SEC));
+  printf("time_pdf: %f %%\n", bt->pdf / bt->main * 100.);
+  printf("time_pdf_mupdf: %f %%\n", bt->pdf_mupdf / bt->main * 100.);
+  printf("time_pdf_findplots: %f %%\n", bt->pdf_findplots / bt->main * 100.);
+  printf("time_pdf_loopplots: %f %%\n", bt->pdf_loopplots / bt->main * 100.);
+  printf("time_pdf_dct: %f %%\n", bt->pdf_dct / bt->main * 100.);
 
 }
